Change Chainable to use *args and **kwargs
This change has some nice features. It is now possible and recommended
to use Link and Chain as normal functions with arguments or keyword
arguments. For backwards compatibility reasons it is still possible to
call the Chainable with a dictionary containing the arguments.

Further the a Link or Chain can now return its output and the complete
dictionary as a tuple (Chainable.full(*args, **kwargs))

Chainable will raise a TypeError if args is not empty and the number of
arguments does not match or a dict is passed as first argument together
with other arguments (This might be problematic?)

Chainable will raise a KeyError if some keys are missing in the input.